# RFC2616 bug fix and performance improvement