--- a/rfcs/rFC2616.ml
+++ b/rfcs/rFC2616.ml
@@ -4,32 +4,29 @@ module P = struct
   let is_space =
     function | ' ' | '\t' -> true | _ -> false
 
-  let is_eol c = c = '\r' || c = '\n'
+  let is_eol =
+    function | '\r' | '\n' -> true | _ -> false
 
   let is_hex =
     function | '0' .. '9' | 'a' .. 'f' | 'A' .. 'F' -> true | _ -> false
-
-  let is_colon c = c = ':'
-
-  let is_colon_or_space c =
-    is_space c || is_colon c
 
   let is_digit =
     function '0' .. '9' -> true | _ -> false
 
   let is_separator =
-    let ss = "()<>@,;:\\\"/[]?={} \t" in
-    let len = String.length ss in
-    let rec loop c i =
-      if i >= len then false
-      else if c = String.unsafe_get ss i then true
-      else loop c (i + 1)
-    in
-    fun c -> loop c 0
+    function
+      | ')' | '(' | '<' | '>' | '@' | ',' | ';' | ':' | '\\' | '"'
+      | '/' | '[' | ']' | '?' | '=' | '{' | '}' | ' ' | '\t' -> true
+      | _ -> false
 
-  let is_token c =
-    let i = Char.code c in
-    i > 31 && i <> 127 && not (is_separator c)
+  let is_token =
+    (* The commented-out ' ' and '\t' are not necessary because of the range at
+     * the top of the match. *)
+    function
+      | '\000' .. '\031' | '\127'
+      | ')' | '(' | '<' | '>' | '@' | ',' | ';' | ':' | '\\' | '"'
+      | '/' | '[' | ']' | '?' | '=' | '{' | '}' (* | ' ' | '\t' *) -> false
+      | _ -> true
 end
 
 let token = take_while1 P.is_token