This pull request includes a bugfix in the response RFC2616 response parser, as well as a performance improvement to various components of the RFC2616 parser. Microbenchmarks suggest a 30% performance improvement for request parsing.

Here are the results of the benchmark before the patches:

```
$ ./pure_benchmark.native -q 60 +time
Estimated testing time 2m (2 benchmarks x 1m). Change using -quota SECS.
┌──────┬──────────┬──────────┬───────────────┬─────────┬──────────┬──────────┬────────────┐
│ Name │ Time R^2 │ Time/Run │          95ci │ mWd/Run │ mjWd/Run │ Prom/Run │ Percentage │
├──────┼──────────┼──────────┼───────────────┼─────────┼──────────┼──────────┼────────────┤
│ json │     1.00 │  21.07ms │ -0.55% +0.55% │ 18.78Mw │ 242.08kw │ 242.08kw │     44.79% │
│ http │     1.00 │  47.05ms │ -0.31% +0.36% │ 32.06Mw │  71.01kw │  71.01kw │    100.00% │
└──────┴──────────┴──────────┴───────────────┴─────────┴──────────┴──────────┴────────────┘
```

... and here is the same benchmark after the patches:

```
$ ./pure_benchmark.native -q 60 +time
Estimated testing time 2m (2 benchmarks x 1m). Change using -quota SECS.
┌──────┬──────────┬──────────┬───────────────┬─────────┬──────────┬──────────┬────────────┐
│ Name │ Time R^2 │ Time/Run │          95ci │ mWd/Run │ mjWd/Run │ Prom/Run │ Percentage │
├──────┼──────────┼──────────┼───────────────┼─────────┼──────────┼──────────┼────────────┤
│ json │     1.00 │  21.17ms │ -0.50% +0.51% │ 18.78Mw │ 242.08kw │ 242.08kw │     58.54% │
│ http │     1.00 │  36.16ms │ -0.33% +0.35% │ 32.06Mw │  71.00kw │  71.00kw │    100.00% │
└──────┴──────────┴──────────┴───────────────┴─────────┴──────────┴──────────┴────────────┘
```